Average Contract Difficulty Asdjusting
- added functionality to adjust the average contract difficulty based
on the contracts the player turns in and the ones they let expire in
their active list

+++ Assets/Scripts/SaveableData.cs
@@ -9,7 +9,7 @@ public class SaveableData
 	public List<LumberContract> activeContracts;
 	public List<LumberContract> availableContracts;
 
-	public int averageContractDifficulty;
+	public float averageContractDifficulty;
 	public List<int> pastGeneratedContractDifficulties;
 
 	public int currentEnergy;
